requirements.txt
fastapi
uvicorn
pydantic
pymongo
PyJWT
